requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-cloud-texttospeech==2.14.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.2
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
python-jose==3.3.0
bcrypt==4.0.1
structlog==23.2.0
slowapi==0.1.8
langdetect==1.0.9
prometheus-client==0.19.0
sentry-sdk==1.35.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
